requirements.txt
alembic==1.0.2
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
factory-boy==2.11.1
Faker==0.9.2
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.6
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
inflect==1.0.1
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
PyJWT==1.6.4
pytest==3.10.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.13
termcolor==1.1.0
text-unidecode==1.2
Werkzeug==0.14.1
